PROD-13 styles applied, animation and background
