+++ src/App.jsx
@@ -1,11 +1,78 @@
 import React from 'react';
 import './App.css';
+import './App.scss';
+import Header from './Components/Header/Header.jsx';
+import sticky from './Functions/Sticky';
 
 function App() {
     const name = 'Martynas';
+    window.onscroll = () => sticky();
     return (
         <div className="App">
-            <h1>Hello World</h1>
+            <Header />
+            <main>
+                <h1>Hello World</h1>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+                <p>
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
+                    Recusandae, aspernatur excepturi itaque consequatur veniam officia impedit eius ducimus
+                    doloremque aut beatae, quisquam velit provident sapiente natus sit quae alias sequi! 12000
+                </p>
+            </main>
         </div>
     );
 }
