+++ src/App.scss
@@ -1,0 +1,9 @@
+.sticky + main {
+    padding-top: 150px;
+    @media screen and (max-width: 912px) {
+        padding-top: 130px;
+      }
+      @media screen and (max-width: 540px) {
+        padding-top: 100px;
+      }
+}
