+++ src/Components/Header/Header.scss
@@ -1,0 +1,171 @@
+@import url('https://fonts.googleapis.com/css2?family=Bitter:wght@800&family=Bitter:wght@500&family=Kalam&family=Kaushan+Script&display=swap');
+
+.header {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  height: auto;
+  background-image: url(../../../public/img/header-darken.png);
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center 44%;
+  transition: all 1s;
+  -webkit-transition: all 1s;
+  -moz-transition: all 1s;
+  -ms-transition: all 1s;
+  -o-transition: all 1s;
+  padding-top: 3%;
+  padding-bottom: 5%;
+  color: #fff;
+
+  .header-title {
+    line-height: 0%;
+    text-align: center;
+    margin: 0 auto;
+    font-family: 'Bitter', serif;
+    font-size: 1rem;
+
+    .header-title--desc {
+      line-height: 80%;
+    }
+
+    h1 {
+      font-size: 2rem;
+
+      span {
+        font-family: 'Kalam', cursive;
+        font-size: 2.4rem;
+        font-weight: 300;
+      }
+    }
+
+    hr {
+      width: 25%;
+      background-color: #fff;
+    }
+  }
+
+  .header-logo {
+    margin-top: 0.1rem;
+    margin-left: 10%;
+    line-height: 100%;
+
+    svg {
+      transition: all 1s;
+      -webkit-transition: all 1s;
+      -moz-transition: all 1s;
+      -ms-transition: all 1s;
+      -o-transition: all 1s;
+      width: 35px;
+      height: 35px;
+      fill: #fff;
+      cursor: pointer;
+    }
+
+    svg:hover {
+      transition: transform 1s;
+      -webkit-transition: transform 1s;
+      -moz-transition: transform 1s;
+      -ms-transition: transform 1s;
+      -o-transition: transform 1s;
+      transform: rotate(30deg);
+    }
+  }
+
+  .header-nav {
+    margin-top: 0.1rem;
+    margin-right: 10%;
+    line-height: 100%;
+    height: 2rem;
+
+    a {
+      text-decoration: none;
+      color: #fff;
+      font-size: 2rem;
+    }
+  }
+
+  @media screen and (max-width: 912px) {
+    background-position: center 41%;
+
+    //height: 10vh;
+    .header-title {
+      font-size: 1rem;
+
+      h1 {
+        font-size: 1.7rem;
+
+        span {
+          font-size: 2rem;
+        }
+      }
+
+      hr {
+        width: 35%;
+      }
+    }
+  }
+
+  @media screen and (max-width: 540px) {
+    background-position: center 40%;
+
+    .header-title {
+      font-size: 0.6rem;
+
+      h1 {
+        font-size: 1.2rem;
+
+        span {
+          font-size: 1.4rem;
+        }
+      }
+
+      hr {
+        width: 35%;
+      }
+    }
+
+    .header-logo {
+      margin-left: 5%;
+
+      svg {
+        width: 25px;
+        height: 25px;
+      }
+    }
+
+    .header-nav {
+      margin-right: 5%;
+      height: 1.5rem;
+
+      a {
+        font-size: 1.5rem;
+      }
+    }
+  }
+}
+
+.sticky {
+  overflow: hidden;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  background-position: center 59%;
+  transition: all 1s;
+  -webkit-transition: all 1s;
+  -moz-transition: all 1s;
+  -ms-transition: all 1s;
+  -o-transition: all 1s;
+  padding-top: 3%;
+  padding-bottom: 2%;
+
+  @media screen and (max-width: 912px) {
+    background-position: center 61%;
+  }
+
+  @media screen and (max-width: 540px) {
+    background-position: center 61%;
+  }
+}
